IPC: Add a proxy interface for Dual-M3 with SMP support
Add the specific interrupt functions for handling the
mbox interrupts for Dual-M3 with SMP support.
This functions are defined in InterruptIpu and they are
plugged in through the generic Interrupt Proxy layer.

The platform differentiation (SMP, non-SM) is resolved at
build time by package.bld, based on if the SMP environment
variable is defiend or not.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>